requirements.txt
django==1.6
pillow
djangorestframework
markdown
django-rest-swagger